Refactor components. Move from lib.rs to region.rs
