+++ valence_anvil/src/lib.rs
@@ -1,9 +1,8 @@
 use std::collections::BTreeMap;
-use std::fmt::{Debug, Formatter, Result as FmtResult};
-use std::path::{Path, PathBuf};
+use std::fmt::Debug;
+use std::path::PathBuf;
 
-use byteorder::{BigEndian, ByteOrder};
-use region::Region;
+use region::{ChunkTimestamp, Region, RegionPos};
 use tokio::fs::File;
 use tokio::sync::{Mutex, MutexGuard};
 use valence::biome::{Biome, BiomeId};
@@ -180,95 +179,5 @@ impl AnvilWorld {
                 }
             })
             .as_mut())
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialOrd, PartialEq, Eq, Ord)]
-pub struct RegionPos {
-    x: i32,
-    z: i32,
-}
-
-impl From<ChunkPos> for RegionPos {
-    fn from(pos: ChunkPos) -> Self {
-        Self {
-            x: pos.x >> 5,
-            z: pos.z >> 5,
-        }
-    }
-}
-
-impl RegionPos {
-    pub fn path(self, world_root: impl AsRef<Path>) -> PathBuf {
-        world_root
-            .as_ref()
-            .join("region")
-            .join(format!("r.{}.{}.mca", self.x, self.z))
-    }
-
-    pub fn contains(self, chunk_pos: ChunkPos) -> bool {
-        Self::from(chunk_pos) == self
-    }
-}
-
-/// The location of the chunk inside the region file.
-#[derive(Copy, Clone, Debug)]
-struct ChunkSeekLocation {
-    offset_sectors: u32,
-    len_sectors: u8,
-}
-
-impl ChunkSeekLocation {
-    const fn zero() -> Self {
-        Self {
-            offset_sectors: 0,
-            len_sectors: 0,
-        }
-    }
-
-    const fn offset(&self) -> u64 {
-        self.offset_sectors as u64 * 1024 * 4
-    }
-
-    const fn len(&self) -> usize {
-        self.len_sectors as usize * 1024 * 4
-    }
-
-    fn load(&mut self, chunk: [u8; 4]) {
-        self.offset_sectors = BigEndian::read_u24(&chunk[..3]);
-        self.len_sectors = chunk[3];
-    }
-}
-
-/// The timestamp when the chunk was last modified in seconds since epoch.
-#[derive(Copy, Clone)]
-pub struct ChunkTimestamp(u32);
-
-impl Debug for ChunkTimestamp {
-    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
-        write!(f, "{}s", self.0)
-    }
-}
-
-impl ChunkTimestamp {
-    const fn zero() -> Self {
-        Self(0)
-    }
-
-    fn load(&mut self, chunk: [u8; 4]) {
-        self.0 = BigEndian::read_u32(&chunk)
-    }
-
-    fn into_option(self) -> Option<Self> {
-        if self.0 == 0 {
-            None
-        } else {
-            Some(self)
-        }
-    }
-
-    #[inline(always)]
-    pub fn seconds_since_epoch(self) -> u32 {
-        self.0
     }
 }
